+++ drivers/char/tpm/tpm_tis_spi_main.c
@@ -318,10 +318,10 @@ static void tpm_tis_spi_remove(struct spi_device *dev)
 }
 
 static const struct spi_device_id tpm_tis_spi_id[] = {
-	{ "st33htpm-spi", (unsigned long)tpm_tis_spi_probe },
-	{ "slb9670", (unsigned long)tpm_tis_spi_probe },
-	{ "tpm_tis_spi", (unsigned long)tpm_tis_spi_probe },
-	{ "tpm_tis-spi", (unsigned long)tpm_tis_spi_probe },
+	{ "st33htpm-spi", (uintptr_t)tpm_tis_spi_probe },
+	{ "slb9670", (uintptr_t)tpm_tis_spi_probe },
+	{ "tpm_tis_spi", (uintptr_t)tpm_tis_spi_probe },
+	{ "tpm_tis-spi", (uintptr_t)tpm_tis_spi_probe },
 	{ "cr50", (unsigned long)cr50_spi_probe },
 	{}
 };
