+++ drivers/gpu/drm/amd/amdgpu/amdgpu_drv.c
@@ -2197,7 +2197,7 @@ static int amdgpu_pci_probe(struct pci_dev *pdev,
 {
 	struct drm_device *ddev;
 	struct amdgpu_device *adev;
-	unsigned long flags = ent->driver_data;
+	uintptr_t flags = ent->driver_data;
 	int ret, retry = 0, i;
 	bool supports_atomic = false;
