+++ drivers/usb/host/xhci-pci.c
@@ -888,10 +888,10 @@ static const struct xhci_driver_data reneses_data = {
 /* PCI driver selection metadata; PCI hotplugging uses this */
 static const struct pci_device_id pci_ids[] = {
 	{ PCI_DEVICE(PCI_VENDOR_ID_RENESAS, 0x0014),
-		.driver_data =  (unsigned long)&reneses_data,
+		.driver_data =  (uintptr_t)&reneses_data,
 	},
 	{ PCI_DEVICE(PCI_VENDOR_ID_RENESAS, 0x0015),
-		.driver_data =  (unsigned long)&reneses_data,
+		.driver_data =  (uintptr_t)&reneses_data,
 	},
 	/* handle any USB 3.0 xHCI controller */
 	{ PCI_DEVICE_CLASS(PCI_CLASS_SERIAL_USB_XHCI, ~0),
