+++ sound/soc/sof/amd/pci-acp63.c
@@ -91,7 +91,7 @@ static void acp63_pci_remove(struct pci_dev *pci)
 /* PCI IDs */
 static const struct pci_device_id acp63_pci_ids[] = {
 	{ PCI_DEVICE(PCI_VENDOR_ID_AMD, ACP_PCI_DEV_ID),
-	.driver_data = (unsigned long)&acp63_desc},
+	.driver_data = (uintptr_t)&acp63_desc},
 	{ 0, }
 };
 MODULE_DEVICE_TABLE(pci, acp63_pci_ids);
